+++ src/2d/ScrollArea.h
@@ -20,7 +20,7 @@ namespace ReyEngine {
 
         void hideVSlider(bool hidden);
         void hideHSlider(bool hidden);
-        Pos<int> getViewportSize(){return _viewport.size();};
+        Pos<double> getViewportSize(){return _viewport.size();};
     protected:
         static constexpr std::string_view VSLIDER_NAME = "__vslider";
         static constexpr std::string_view HSLIDER_NAME = "__hslider";
@@ -36,17 +36,17 @@ namespace ReyEngine {
         void _on_child_added(std::shared_ptr<BaseWidget> &child) override;
         void _init() override;
         Handled _unhandled_input (const InputEvent&, const std::optional<UnhandledMouseInput>&) override;
-        std::optional<std::shared_ptr<BaseWidget>> askHover(const ReyEngine::Pos<int>& globalPos) override;
+        std::optional<std::shared_ptr<BaseWidget>> askHover(const ReyEngine::Pos<double>& globalPos) override;
 //        Handled _unhandled_masked_input(const InputEventMouse&, const std::optional<UnhandledMouseInput>&) override;
         ///Ignores scroll bars, but adds their widths if they are visible
-        ReyEngine::Size<int> getScrollAreaChildBoundingBox();
+        ReyEngine::Size<double> getScrollAreaChildBoundingBox();
 
         ReyEngine::Range<double> scrollOffsetX;
         ReyEngine::Range<double> scrollOffsetY;
         std::shared_ptr<Slider> vslider;
         std::shared_ptr<Slider> hslider;
-        ReyEngine::Size<int> _childBoundingBox;
-        Rect<int> _viewport; //the area that isn't obscured by sliders
+        ReyEngine::Size<double> _childBoundingBox;
+        Rect<double> _viewport; //the area that isn't obscured by sliders
         bool _hideVSlider = false;
         bool _hideHSlider = false;
     };
