+++ src/2d/Sprite.h
@@ -10,7 +10,7 @@ namespace ReyEngine{
    REYENGINE_OBJECT_BUILD_ONLY(Sprite, BaseWidget)
    , PROPERTY_DECLARE(region)
    {}
-   static std::shared_ptr<Sprite> build(const std::string& name, const FileSystem::File& texPath, const Rect<int>& region){
+   static std::shared_ptr<Sprite> build(const std::string& name, const FileSystem::File& texPath, const Rect<double>& region){
       auto me = build(name);
       me->texPath = texPath;
       me->region = region;
@@ -23,13 +23,13 @@ namespace ReyEngine{
       void _init() override;
       bool setTexture(const FileSystem::File&);
       std::optional<const std::reference_wrapper<ReyTexture>> getTexture();
-      void setRegion(const Rect<int>&);
+      void setRegion(const Rect<double>&);
       void fitTexture(); //makes the sprite the same size as the texture REGION (not necessarily the texture SIZE!)
 
    protected:
       FileSystem::File texPath;
       std::unique_ptr<ReyTexture> texture;
-      RectProperty<int> region;
+      RectProperty<double> region;
 
    private:
       bool _fitNextTexture = false; //if we dont' yet have a texture (because window isn't loaded), then fit the texture when we do have one
